requirements.txt
async-generator==1.10
attrs==23.1.0
autopep8==2.0.0
blinker==1.6.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coverage==7.2.5
exceptiongroup==1.1.1
filelock==3.10.7
Flask==2.3.2
h11==0.14.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
more-itertools==9.1.0
numpy==1.25.1
outcome==1.2.0
packaging==23.1
pipdeptree==2.10.2
pluggy==1.0.0
pycodestyle==2.9.1
pycparser==2.21
PySocks==1.7.1
pytest-cov==4.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
qase-pytest==5.0.1
qase-python-commons==2.0.2
qaseio==3.3.0
requests==2.29.0
selenium==4.9.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
urllib3==1.26.15
webdriver-manager==3.8.6
Werkzeug==2.3.6
wsproto==1.2.0
